Add connection string feature
This way users will first have to send the connection string, before
being allowed to interact or even be recognized as connected in kchat.

It has the minor inconvenience of requiring supporting the connection
string lines, and can be disabled by virtue of a compilation flag,
the string being given to the user can also be disabled.

Signed-off-by: Siren <siren@kernal.eu>